+++ isochrones.go
@@ -11,16 +11,16 @@ import (
 // See ref. https://wiki.openstreetmap.org/wiki/Isochrone and https://en.wikipedia.org/wiki/Isochrone_map
 // Note: implemented breadth-first searching path algorithm does not guarantee shortest pathes to reachable vertices (until all edges have cost 1.0). See ref: https://en.wikipedia.org/wiki/Breadth-first_search
 // Note: result for estimated costs could be also inconsistent due nature of data structure
-func (graph *Graph) Isochrones(source int64, maxCost float64) (map[int64]float64, error) {
+func (graph *Graph) Isochrones(source int, maxCost float64) (map[int]float64, error) {
 	ok := true
 	if source, ok = graph.mapping[source]; !ok {
 		return nil, fmt.Errorf("No such source")
 	}
 	Q := &minheapSTD{}
 	heap.Init(Q)
-	distance := make(map[int64]float64, len(graph.Vertices))
+	distance := make(map[int]float64, len(graph.Vertices))
 	Q.Push(minheapNode{id: source, distance: 0})
-	visit := make(map[int64]bool)
+	visit := make(map[int]bool)
 	for Q.Len() != 0 {
 		next := heap.Pop(Q).(minheapNode)
 		visit[next.id] = true
@@ -32,7 +32,7 @@ func (graph *Graph) Isochrones(source int64, maxCost float64) (map[int64]float64
 				neighbor := vertexList[i]
 				if v1, ok1 := graph.shortcuts[next.id]; ok1 {
 					if _, ok2 := v1[neighbor]; ok2 {
-						// Ignore contract
+						// Ignore shortuct
 						continue
 					}
 				}
